requirements.txt
Django==4.1.9
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==3.2.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
PyJWT==2.7.0
PyMuPDF==1.20.2
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
requests==2.30.0
PyJWT==2.7.0
whitenoise==6.5.0